remove modeling_fast; move config into config file; format code
