+++ paddleformers/examples/deepseek_v3/config/pretrain_argument.json
@@ -1,5 +1,5 @@
 {
-    "model_name_or_path": "DeepseekV2ModelFast",
+    "model_name_or_path": "./config/",
     "tokenizer_name_or_path": "deepseek-ai/DeepSeek-V3",
     "input_dir": "./data",
     "output_dir": "./checkpoints/pretrain_ckpts",
